move update owner fn to utils
