+++ typescript/infra/src/utils/safe.ts
@@ -6,7 +6,7 @@ import {
   SafeTransaction,
 } from '@safe-global/safe-core-sdk-types';
 import chalk from 'chalk';
-import { ethers } from 'ethers';
+import { BigNumber, ethers } from 'ethers';
 
 import {
   ChainNameOrId,
@@ -14,7 +14,10 @@ import {
   getSafe,
   getSafeService,
 } from '@hyperlane-xyz/sdk';
-import { Address, CallData } from '@hyperlane-xyz/utils';
+import { Address, CallData, eqAddress } from '@hyperlane-xyz/utils';
+
+import safeSigners from '../../config/environments/mainnet3/safe/safeSigners.json' assert { type: 'json' };
+import { AnnotatedCallData } from '../govern/HyperlaneAppGovernor.js';
 
 export async function getSafeAndService(
   chain: ChainNameOrId,
@@ -221,4 +224,51 @@ export async function deleteSafeTx(
       error,
     );
   }
+}
+
+export async function updateSafeOwner(
+  safeSdk: Safe.default,
+): Promise<AnnotatedCallData[]> {
+  const threshold = await safeSdk.getThreshold();
+  const owners = await safeSdk.getOwners();
+  const newOwners = safeSigners.signers;
+  const ownersToRemove = owners.filter(
+    (owner) => !newOwners.some((newOwner) => eqAddress(owner, newOwner)),
+  );
+  const ownersToAdd = newOwners.filter(
+    (newOwner) => !owners.some((owner) => eqAddress(newOwner, owner)),
+  );
+
+  console.log(chalk.magentaBright('Owners to remove:', ownersToRemove));
+  console.log(chalk.magentaBright('Owners to add:', ownersToAdd));
+
+  const transactions: AnnotatedCallData[] = [];
+
+  for (const ownerToRemove of ownersToRemove) {
+    const { data: removeTxData } = await safeSdk.createRemoveOwnerTx({
+      ownerAddress: ownerToRemove,
+      threshold,
+    });
+    transactions.push({
+      to: removeTxData.to,
+      data: removeTxData.data,
+      value: BigNumber.from(removeTxData.value),
+      description: `Remove safe owner ${ownerToRemove}`,
+    });
+  }
+
+  for (const ownerToAdd of ownersToAdd) {
+    const { data: addTxData } = await safeSdk.createAddOwnerTx({
+      ownerAddress: ownerToAdd,
+      threshold,
+    });
+    transactions.push({
+      to: addTxData.to,
+      data: addTxData.data,
+      value: BigNumber.from(addTxData.value),
+      description: `Add safe owner ${ownerToAdd}`,
+    });
+  }
+
+  return transactions;
 }
